requirements.txt
hopsworks
joblib
scikit-learn
seaborn
dataframe-image
modal-client
gradio
yfinance